refactor / cleanup code
