+++ internal/protocols/protocol.go
@@ -1,10 +1,22 @@
 package protocols
 
 import (
+	"bufio"
 	"github.com/404minds/avl-receiver/internal/protocols/fm1200"
 	"github.com/404minds/avl-receiver/internal/protocols/gt06"
+	"io"
+
 	"github.com/404minds/avl-receiver/internal/types"
 )
+
+type DeviceProtocol interface {
+	GetDeviceType() types.DeviceType
+	SetDeviceType(types.DeviceType)
+	GetProtocolType() types.DeviceProtocolType
+	GetDeviceID() string
+	Login(*bufio.Reader) ([]byte, int, error)
+	ConsumeStream(*bufio.Reader, io.Writer, chan types.DeviceStatus) error
+}
 
 func MakeProtocolForType(t types.DeviceProtocolType) DeviceProtocol {
 	switch t {
